+++ src/constants/filters.ts
@@ -1,0 +1,3 @@
+export const CATEGORIES = ['게임', '영화', '맛집탐방', '운동', '스터디', '기타'];
+
+export const RADII = ['1km', '3km', '5km'];
